+++ src/main/java/com/anthonyhilyard/iceberg/mixin/MinecraftMixin.java
@@ -68,7 +68,6 @@ public class MinecraftMixin
 					Loader.LOGGER.debug(ExceptionUtils.getStackTrace(e));
 				}
 			}
-
 		}
 		renderTarget.clear(onOSX);
 	}